added validation to migration

+++ db/migrate/20201014193527_update_rails_lti2_provider_tool_index.rb
@@ -3,7 +3,7 @@
 class UpdateRailsLti2ProviderToolIndex < ActiveRecord::Migration[6.0]
   def self.up
     add_index('rails_lti2_provider_tools', ['uuid'], name: 'index_uuid', unique: true)
-    remove_index('rails_lti2_provider_tools', %w[id tenant_id])
+    remove_index('rails_lti2_provider_tools', %w[id tenant_id]) if index_exists?(:id, :tenant_id)
   end
 
   def self.down
